Products index view updated

+++ database/migrations/2025_05_28_151920_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name',255);
             $table->decimal('price', 10, 2); 
             $table->string('category');
-            $table->bigInteger('discount')->nullable(); 
+            $table->decimal('discount', 8, 2)->nullable();
             $table->text('gallery')->nullable(); 
             $table->text('description')->nullable();
             $table->timestamps();
